+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.3.0
+* Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
+* Coinbase is no longer a supported payment method. `PaymentMethodNoLongerSupported` will be returned for Coinbase operations.
+* Add facilitated details to Transaction if present
+* Add `submit_for_settlement` to `Subscription.retryCharge`
+* Add `options` -> `paypal` -> `description` for creating and updating subscriptions
+* Add Dispute API
+* Add DocumentUpload API
+* Add `deviceDataCaptured` field in `riskData`
+* Add support for upgrading a PayPal future payment refresh token to a billing agreement
+
 ## 2.2.0
 * Fixes wrap-promise issue (closes #102) - Thanks @targunp
 * Add iDEAL to PaymentInstrumentTypes